### Jira ticket url
- https://zuelligpharma.atlassian.net/browse/EZRX-10741

### TDD PR url
-

### Root cause
- As per the current implementation in web when a bonus item is added more than once, after revisiting the add bonus bottomsheet, it is treated as a new material.

### Solution
- As a solution have considered the itemID to determine. weather the item has already been added to cart or not as a result of which a bonus item is considered as a new material only if the itemId is not present inside the BonusMaterialState.

### Test account
- environment:UAT
- username:rootadmin
- salesOrg:2001
- customer:any
- shipto:any
- material/setup:

### Output screenshot
-![Simulator Screen Shot - iPhone 14 Pro Max - 2023-09-08 at 16.43.46.png](https://dev.azure.com/zuelligpharmadevops/3368559d-402f-4ca8-a029-737bca870679/_apis/git/repositories/d10de74c-8ac2-4e8e-af4b-a023b6c9dee2/pullRequests/77280/attachments/Simulator%20Screen%20Shot%20-%20iPhone%2014%20Pro%20Max%20-%202023-09-08%20at%2016.43.46.png =x500)

### Reminder
- if you code have significant change please run integration test `make run_sg_client_test` to check if you break any test cases

EZRX-10741 Mobile v3: Need to align with web on bonus sample item added to cart

+++ lib/application/order/cart/cart_bloc.dart
@@ -142,7 +142,7 @@ class CartBloc extends Bloc<CartEvent, CartState> {
           salesOrganisationConfig: state.config,
           shipToInfo: state.shipToInfo,
           materialInfo: e.bonusMaterial,
-          quantity: e.bonusMaterial.quantity,
+          quantity: e.bonusMaterial.quantity.intValue,
           language: e.user.settings.languagePreference.languageCode,
           counterOfferDetails: e.counterOfferDetails,
           itemId: e.bonusItemId.getValue(),
@@ -515,7 +515,7 @@ class CartBloc extends Bloc<CartEvent, CartState> {
                   (element) => !element.materialInfo.type.typeBundle
                       ? PriceAggregate.empty().copyWith(
                           materialInfo: element.materialInfo,
-                          quantity: element.materialInfo.quantity,
+                          quantity: element.materialInfo.quantity.intValue,
                           salesOrgConfig: element.salesOrgConfig,
                           bonusSampleItems: element.bonusSampleItems,
                         )
